changed: name of the user collection and also of the tasks, fixed: error when inserting date into database through moment library
