+++ server/controllers/taskController.js
@@ -1,12 +1,13 @@
-import PostTask from '../models/postTask.js'
-import PostUser from '../models/postUser.js'
+import TaskSchema from '../models/taskSchema.js'
+import UserSchema from '../models/userSchema.js'
 import moment from 'moment';
 
 const dataAtual = moment().format('DD-MM-YYYY');
 const horaAtual = moment().format('LTS');
+const currMoment = `${horaAtual} ${dataAtual}`;
 
 export const getAllTasks = async (req, res) => {
-    const getTasks = await PostTask.find();
+    const getTasks = await TaskSchema.find();
 
     if (!getTasks) return res.status(204).json({ message: 'user have not any task yet' });
     return res.status(200).json(getTasks);
@@ -14,8 +15,8 @@ export const getAllTasks = async (req, res) => {
 
 export const getTasks = async (req, res) => {
   const { id } = req.body;
-    const getUser = await PostUser.findOne({ _id: id });
-    const getTasks = await PostTask.find({ autor: id });
+    const getUser = await UserSchema.findOne({ _id: id });
+    const getTasks = await TaskSchema.find({ autor: id });
 
     if (!getTasks) return res.status(204).json({ message: 'user have not any task yet' });
     return res.status(200).json({ user: getUser, tasks: getTasks});
@@ -24,7 +25,7 @@ export const getTasks = async (req, res) => {
 export const addTask = async (req, res) => {
   const data = req.body;
   const { email, id, prioridade, nome, descricao } = data;
-  const newUser = new PostTask({autor: id, prioridade, nome, descricao });
+  const newUser = new TaskSchema({autor: id, prioridade, nome, dataDeCriacao: currMoment, descricao });
   try {
     const added = await newUser.save();
     res.status(200).json(added);
@@ -37,7 +38,7 @@ export const initTask = async (req, res) => {
   const { id } = req.params;
 
   try {
-    await PostTask.updateOne({_id: id }, { status: 'Em andamento' });
+    await TaskSchema.updateOne({_id: id }, { status: 'Em andamento' });
 
     res.status(200).json({update: true});
   } catch (error) {
@@ -49,7 +50,7 @@ export const doneTask = async (req, res) => {
   const { id } = req.params;
 
   try {
-    await PostTask.updateOne({_id: id }, { status: 'Pronto', dataDeConclusao: `${dataAtual - horaAtual}` });
+    await TaskSchema.updateOne({_id: id }, { status: 'Pronto', dataDeConclusao: currMoment });
 
     res.status(200).json({done: true});
   } catch (error) {
