+++ server/models/taskSchema.js
@@ -6,26 +6,23 @@ const horaAtual = moment().format('LTS');
 
 // src: https://mongoosejs.com/docs/populate.html
 const taskSchema = mongoose.Schema({
-  autor: { type: mongoose.Schema.Types.ObjectId, ref: 'PostUser' },
+  autor: { type: mongoose.Schema.Types.ObjectId, ref: 'UserSchema' },
   prioridade: String,
   nome: String,
   descricao: String,
-  dataDeCriacao: {
-    type: String,
-    default: `${dataAtual - horaAtual}`
-  },
+  dataDeCriacao: String,
   status: {
     type: String,
     default: 'Pendente'
   },
   dataDeConclusao: {
-    type: Date,
+    type: String,
     default: null
   }
 },
 // https://stackoverflow.com/questions/13699784/mongoose-v-property-hide
 { versionKey: false });
 
-const TaskSchema = mongoose.model('taskSchema', taskSchema);
+const TaskSchema = mongoose.model('TaskSchema', taskSchema);
 
 export default TaskSchema;
